fix name conflict with Arduino/Time.h
